Add HTTP client implementation
This introduces package rest

The idea is to create an abstraction layer over the Resty client to make
the codebase agnostic to the underlying HTTP client.
This involves defining an interface for the HTTP client and a wrapper
for the Resty client that implements this interface.
This way, Resty can be easily swapped out for another HTTP client in
the future without affecting the calling functions.

Signed-off-by: Raamsri Kumar <raam@tinkershack.in>